backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
onnxruntime
numpy
Pillow

